DAOTHER-8063 moved upload functions to uploadProvider
