+++ azure-pipelines.yml
@@ -15,36 +15,23 @@ trigger:
 pr:
   - master
 
-schedules:
-  - cron: "0 12 * * 0"
-    displayName: Weekly Sunday build
-    branches:
-      include:
-        - master
-    always: true
+#schedules:
+#  - cron: "0 12 * * 0"
+#    displayName: Weekly Sunday build
+#    branches:
+#      include:
+#        - master
+#    always: true
 
 jobs:
-#  - template: azure/azure-template-tox-job.yml
-#    parameters: { tox: "format_pep8", python: "3.9", os: "linux" }
-#  - template: azure/azure-template-tox-job.yml
-#    parameters: { tox: "lint_pylint", python: "3.9", os: "linux" }
-  #
-#  - template: azure/azure-template-tox-job.yml
-#    parameters: { tox: "test_coverage", python: "3.9", os: "linux" }
-  #
   - template: azure/azure-template-tox-job.yml
     parameters: { tox: "py39", python: "3.9", os: "linux" }
   #
-  #- template: azure-template-tox-job.yml
-  #  parameters: {tox: 'py39', python: '3.9', os: 'macos'}
-  #- template: azure-template-publish-job.yml
-  #  parameters: {tox: 'py39', python: '3.9', os: 'macos'}
+#  - template: azure/azure-template-publish-job.yml
+#    parameters: { tox: "py39", python: "3.9", os: "linux" }
   #
-  - template: azure/azure-template-publish-job.yml
-    parameters: { tox: "py39", python: "3.9", os: "linux" }
-  #
-  - template: azure/azure-template-docker-job.yml
-    parameters: { os: "linux", docker_task: "build_and_test" }
-  - template: azure/azure-template-docker-publish-job.yml
-    parameters: { os: "linux", docker_task: "build_and_publish" }
+#  - template: azure/azure-template-docker-job.yml
+#    parameters: { os: "linux", docker_task: "build_and_test" }
+#  - template: azure/azure-template-docker-publish-job.yml
+#    parameters: { os: "linux", docker_task: "build_and_publish" }
 #
