+++ tox.ini
@@ -31,6 +31,6 @@ commands =
     echo " ------- Consolidating {envname} data ----------"
     coverage combine
     echo " ------- Building {envname} reports ----------"
-    coverage report --fail-under=65
+    coverage report --fail-under=60
     - coverage xml
     echo "Completed {envname}"
